agents/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Test and demo for Azure AI Foundry Python SDK"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai==2.30.0",
    "azure-ai-projects==2.0.1",
    "python-dotenv==1.2.2",
    "azure-identity==1.25.3",
    "jsonref==1.1.0",
    "requests==2.33.1",
    "ipython==9.12.0",
]
